requirements.txt
numpy
python-dotenv
openai
gradio-client
tqdm
prettytable
tenacity